+++ backend/src/main/java/cm/imf/pipeline/entity/OtpCode.java
@@ -31,7 +31,7 @@ public class OtpCode {
     private OffsetDateTime expiresAt;
 
     @Column(name = "attempts_used", nullable = false)
-    private int attemptsUsed;
+    private short attemptsUsed;
 
     @Column(name = "used", nullable = false)
     private boolean used;
